backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-language==2.11.1
google-generativeai==0.3.1
firebase-admin==6.2.0
aiohttp==3.9.1
beautifulsoup4==4.12.2